Detail_project/requirements.txt
streamlit
llama-index
faiss-cpu
sentence-transformers
PyMuPDF
